+++ Services/ISubscriptionCategoryService.cs
@@ -5,15 +5,15 @@ namespace StreamingSubscriptionTrackerAPI.Services
     public interface ISubscriptionCategoryService
     {
         //GET
-        SubscriptionCategoryResponseDTO GetById(int id);
+        SubscriptionCategoryResponseDTO GetById(long id);
         SubscriptionCategoryResponseDTO GetByName(string name);
         List<SubscriptionCategoryResponseDTO> GetAll();
         //POST
         SubscriptionCategoryResponseDTO Create(SubscriptionCategoryRequestDTO subscriptionCategory);
         //PUT
-        SubscriptionCategoryResponseDTO Update(int id, SubscriptionCategoryRequestDTO subscriptionCategory);
+        SubscriptionCategoryResponseDTO Update(long id, SubscriptionCategoryRequestDTO subscriptionCategory);
         //DELETE
-        void Delete(int id);
+        void Delete(long id);
 
     }
 }
